move contrib.render to infer.inspect
