+++ numpyro/__init__.py
@@ -4,8 +4,8 @@
 import logging
 
 from numpyro import compat, diagnostics, distributions, handlers, infer, ops, optim
-from numpyro.contrib.render import render_model
 from numpyro.distributions.distribution import enable_validation, validation_enabled
+from numpyro.infer.inspect import render_model
 import numpyro.patch  # noqa: F401
 from numpyro.primitives import (
     deterministic,
